pyproject.toml
[project]
name = "tg-chat-parser"
version = "2.0.0"
description = "An interactive command-line tool to export chat history from Telegram channels into beautifully formatted .docx Word documents."
authors = [{ name = "Nick Chets", email = "nick@chets.dev" }]
requires-python = ">=3.10"

dependencies = [
    "telethon>=1.34.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "python-docx>=1.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.2.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.4",
]

[project.scripts]
chat-parser = "src.cli:main_cli"

[tool.ruff]
line-length = 120
exclude = ["test_google_auth.py"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "C90"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
asyncio_mode = "auto"
testpaths = [
    "tests",
]
pythonpath = ["."]
